+++ functions.php
@@ -979,26 +979,3 @@ class MizzouHybridBase extends MizzouBlocks {
 
 // Create object.
 new MizzouHybridBase();
-
-// Add Option Pages.
-if ( function_exists( 'acf_add_options_page' ) ) {
-
-	acf_add_options_page(
-		array(
-			'page_title' => 'Theme Settings',
-			'menu_title' => 'Theme Settings',
-			'menu_slug'  => 'theme-general-settings',
-			'capability' => 'manage_options',
-			'redirect'   => false,
-		)
-	);
-
-	acf_add_options_sub_page(
-		array(
-			'page_title'  => 'Footer Settings',
-			'menu_title'  => 'Footer',
-			'parent_slug' => 'theme-general-settings',
-			'capability'  => 'edit_theme_options',
-		)
-	);
-}