+++ inc/acf/definitions/acf-ds-layer-definitions.php
@@ -245,7 +245,7 @@ if ( function_exists( 'acf_add_local_field_group' ) ) :
 			'hide_on_screen'        => '',
 			'active'                => true,
 			'description'           => '',
-			'show_in_rest'          => 0,
+			'show_in_rest'          => 1,
 		)
 	);
 
@@ -394,7 +394,7 @@ if ( function_exists( 'acf_add_local_field_group' ) ) :
 			'hide_on_screen'        => '',
 			'active'                => true,
 			'description'           => '',
-			'show_in_rest'          => 0,
+			'show_in_rest'          => 1,
 		)
 	);
 
@@ -648,7 +648,7 @@ if ( function_exists( 'acf_add_local_field_group' ) ) :
 			'hide_on_screen'        => '',
 			'active'                => true,
 			'description'           => '',
-			'show_in_rest'          => 0,
+			'show_in_rest'          => 1,
 		)
 	);
 
@@ -960,7 +960,7 @@ if ( function_exists( 'acf_add_local_field_group' ) ) :
 			'hide_on_screen'        => '',
 			'active'                => true,
 			'description'           => '',
-			'show_in_rest'          => 0,
+			'show_in_rest'          => 1,
 		)
 	);
 
@@ -1172,7 +1172,7 @@ if ( function_exists( 'acf_add_local_field_group' ) ) :
 			'hide_on_screen'        => '',
 			'active'                => true,
 			'description'           => '',
-			'show_in_rest'          => 0,
+			'show_in_rest'          => 1,
 		)
 	);
 
@@ -1322,7 +1322,7 @@ if ( function_exists( 'acf_add_local_field_group' ) ) :
 			'hide_on_screen'        => '',
 			'active'                => true,
 			'description'           => '',
-			'show_in_rest'          => 0,
+			'show_in_rest'          => 1,
 		)
 	);
 
@@ -1497,7 +1497,7 @@ if ( function_exists( 'acf_add_local_field_group' ) ) :
 			'hide_on_screen'        => '',
 			'active'                => true,
 			'description'           => '',
-			'show_in_rest'          => 0,
+			'show_in_rest'          => 1,
 		)
 	);
 
@@ -1626,7 +1626,7 @@ if ( function_exists( 'acf_add_local_field_group' ) ) :
 			'hide_on_screen'        => '',
 			'active'                => true,
 			'description'           => '',
-			'show_in_rest'          => 0,
+			'show_in_rest'          => 1,
 		)
 	);
 
@@ -1876,7 +1876,7 @@ if ( function_exists( 'acf_add_local_field_group' ) ) :
 			'hide_on_screen'        => '',
 			'active'                => true,
 			'description'           => '',
-			'show_in_rest'          => 0,
+			'show_in_rest'          => 1,
 		)
 	);
 
@@ -2239,7 +2239,7 @@ if ( function_exists( 'acf_add_local_field_group' ) ) :
 			'hide_on_screen'        => '',
 			'active'                => true,
 			'description'           => '',
-			'show_in_rest'          => 0,
+			'show_in_rest'          => 1,
 		)
 	);